+++ Settings/sources/com/google/android/material/color/MaterialColors.java
@@ -6,7 +6,7 @@ import android.util.TypedValue;
 import android.view.View;
 import androidx.core.graphics.ColorUtils;
 import com.google.android.material.resources.MaterialAttributes;
-/* loaded from: classes.dex */
+
 public class MaterialColors {
     public static int getColor(View view, int i) {
         return MaterialAttributes.resolveOrThrow(view, i);
@@ -30,7 +30,7 @@ public class MaterialColors {
     }
 
     public static int layer(int i, int i2, float f) {
-        return layer(i, ColorUtils.setAlphaComponent(i2, Math.round(Color.alpha(i2) * f)));
+        return layer(i, ColorUtils.setAlphaComponent(i2, Math.round(((float) Color.alpha(i2)) * f)));
     }
 
     public static int layer(int i, int i2) {
