+++ Settings/sources/com/google/android/material/math/MathUtils.java
@@ -1,5 +1,5 @@
 package com.google.android.material.math;
-/* loaded from: classes2.dex */
+
 public final class MathUtils {
     public static boolean geq(float f, float f2, float f3) {
         return f + f3 >= f2;
@@ -14,7 +14,7 @@ public final class MathUtils {
     }
 
     public static float dist(float f, float f2, float f3, float f4) {
-        return (float) Math.hypot(f3 - f, f4 - f2);
+        return (float) Math.hypot((double) (f3 - f), (double) (f4 - f2));
     }
 
     public static float distanceToFurthestCorner(float f, float f2, float f3, float f4, float f5, float f6) {
